split AST.h
