+++ include/AST/Expr.h
@@ -1,3 +1,4 @@
+#pragma once
 
 namespace AST {
